refactor: created aditionalconstructedTravelPlan entities for code clarity

+++ app/backend/src/entities/TravelPlans/ConstructedExpandedTravelPlan.entity.cr
@@ -1,8 +1,8 @@
 class ConstructedExpandedTravelPlan
-  property id : Int64
+  property id : Int32
   property travel_stops : Array(ExpandedTravelStop)
 
-  def initialize(@id : Int64, @travel_stops : Array(ExpandedTravelStop))
+  def initialize(@id : Int32, @travel_stops : Array(ExpandedTravelStop))
   end
 
   def to_json(json : JSON::Builder)
